Now we are using Maven, so the folder 'Practica' was changed for 'com.dis.practica1'
