--- Practica/almacen.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<Almacen>
-          <Productos>
-                    <Producto Código="wqw">
-                              <Nombre>wq</Nombre>
-                              <Descripción>wq</Descripción>
-                              <Stock>wqw</Stock>
-                              <Pendientes>wq</Pendientes>
-                              <Localización>
-                                        <Pasillo>qw</Pasillo>
-                                        <Estantería>wq</Estantería>
-                                        <Estante>wq</Estante>
-                              </Localización>
-                    </Producto>
-          </Productos>
-          <Clientes>
-                    <Cliente>
-                              <Nombre>rer</Nombre>
-                              <Apellidos>ere</Apellidos>
-                              <Email>re</Email>
-                              <Telefono>re</Telefono>
-                              <Dirección>
-                                        <Calle>rer</Calle>
-                                        <Número>err</Número>
-                                        <CódigoPostal>re</CódigoPostal>
-                                        <Población>re</Población>
-                                        <País>re</País>
-                              </Dirección>
-                    </Cliente>
-          </Clientes>
-          <Pedidos>
-                    <Pedido>
-                              <Productos>hth</Productos>
-                              <Cantidad>hth</Cantidad>
-                              <Destinatario>th</Destinatario>
-                              <Fecha>th</Fecha>
-                              <Dirección>
-                                        <Calle>th</Calle>
-                                        <Número>ht</Número>
-                                        <CódigoPostal>ht</CódigoPostal>
-                                        <Población>ht</Población>
-                                        <País>ht</País>
-                              </Dirección>
-                    </Pedido>
-          </Pedidos>
-</Almacen>
