+++ com.dis.practica1/src/test/java/com/dis/practica1/AppTest.java
@@ -1,0 +1,20 @@
+package com.dis.practica1;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
